fix warning

+++ ir/opt/proc_cloning.c
@@ -311,7 +311,7 @@ static ir_node *get_irg_arg(ir_graph *irg, size_t pos)
 	/* Search the argument with the number pos.*/
 	for (i = get_irn_n_outs(irg_args) - 1; i >= 0; --i) {
 		ir_node *proj = get_irn_out(irg_args, i);
-		if (pos == get_Proj_proj(proj)) {
+		if ((int)pos == get_Proj_proj(proj)) {
 			if (arg) {
 				/*
 				 * More than one arg node found:
